🚚 move files of inner components
