+++ apps/client/src/Routes/Classes/ClassDetails/Mobile/Components/List/index.tsx
