+++ apps/client/src/Routes/Classes/ClassDetails/Mobile/Components/List/styles.ts
