+++ apps/client/src/Routes/Classes/ClassDetails/Mobile/index.tsx
@@ -3,9 +3,9 @@ import React, { FC } from 'react'
 // import { Banner, CourseInfo, CourseResume, ResumeInfo } from './styles'
 import FullPage from "Components/FullPage"
 import { Course } from 'Api/Courses'
-import Banner from './Banner'
-import CourseInfo from "./CourseInfo"
-import CourseResume from "./Resume"
+import Banner from './Components/Banner'
+import CourseInfo from "./Components/CourseInfo"
+import CourseResume from "./Components/Resume"
 
 import Header from 'Components/Header'
 import Footer from 'Components/Footer'
